feat(i18n): adding adapter to use old and new data for usercard display and intl (#NGXCPTS-56)
