+++ projects/user-card/src/adapters/user-card-adapter.services.ts
@@ -1,0 +1,32 @@
+import { Injectable } from '@angular/core';
+
+import { NgxUserCard } from '../user-card.model';
+import { UserCardDataProviderService } from './user-card-data-provider';
+import type { BadgeColor, UserCardDisplayModel } from './user-card-display.model';
+
+@Injectable({
+    providedIn: 'root'
+})
+export class UserCardAdapterService extends UserCardDataProviderService {
+
+    override buildDisplayableUserCard(user: NgxUserCard): UserCardDisplayModel {
+        return {
+            fullname: this.getFullName(user),
+            initials: user.initials,
+            function: this.getFunction(user).toLocaleLowerCase(),
+            specialty: this.getSpecialty(user),
+            badgeColor: (user?.userCardBadgeColor || 'grey') as BadgeColor,
+            service: user.esoN3Label,
+            type: user.type,
+            phoneNumber: this.formatPhone(user.phone),
+            mobileNumber: this.formatPhone(user.bip || user.mobile),
+            email: user.email?.toLowerCase(),
+            address: this.getUserAddress(user)
+        };
+    }
+
+    override getShortTitle(user: NgxUserCard): string {
+        return user?.shortTitle || '';
+    }
+
+}
