+++ projects/user-card/src/index.ts
@@ -1,4 +1,4 @@
 export * from './user-card.model';
 export * from './user-card.component';
 export { NgxUserCardIntl, provideNgxUserCard } from './providers';
-
+export { UserCardAdapteLegacyService as UserCardAdapterService } from './adapters/user-card-adapter-legacy.services';
